\name{tidycorrgram-package}
\alias{tidycorrgram-package}
\alias{tidycorrgram}
\docType{package}
\title{Tidy Correlation Matrices and ggplot2 Correlograms}
\description{
Compute correlation matrices as tidy data frames and create publication-ready
correlograms with \pkg{ggplot2}.
}
\details{
The main functions are \code{\link{corrgram_data}} for computing tidy
correlation tables and \code{\link{corrgram}} for drawing correlograms.
}
\seealso{
\code{\link{corrgram_data}}, \code{\link{corrgram}}
}
